refactor: flatten api/auth slice onto the standard subpackage template
Collapse the twelve-subpackage auth slice into the layout defined by
docs/ARCHITECTURE.md's Slice Pattern. Moves only — no behavior change.

Key moves:
- api/auth/domain/UserPrincipal -> security/UserPrincipal
  (removes the module-name collision with eventitta-domain)
- api/auth/config/SecurityConfig -> slice root (keeps security/ and
  jwt/ acyclic: SecurityConfig imports both, jwt imports security)
- api/auth/constants/AuthConstants -> slice root
- api/auth/mapper/AuthMapper -> controller/AuthMapper
- api/auth/properties/* fanned out next to their consumers
  (SecurityCorsProperties -> security/, JwtProperties -> jwt/,
  CookieProperties -> cookie/, KakaoWebProperties -> oauth/kakao/)
- api/auth/web/* split by intent: CookieManager -> cookie/,
  KakaoAuthorizationSupport -> oauth/kakao/,
  ClientSessionMetadataResolver -> session/
- api/auth/jwt/{filter,service,util}/* flattened into jwt/

The 2-deep oauth/kakao/ path is intentional: any future provider
(Google, Apple) drops in as a sibling without reshaping the tree.

controller/request and controller/response are deliberately left in
place in this PR. They can only land in domain/auth/dto/ alongside
the existing port/dto + service/dto deviation, which deserves its
own coherent sweep in a follow-up PR.
