+++ eventitta-api/src/main/java/com/eventitta/api/auth/jwt/CustomUserDetailsService.java
@@ -1,6 +1,6 @@
-package com.eventitta.api.auth.jwt.service;
+package com.eventitta.api.auth.jwt;
 
-import com.eventitta.api.auth.domain.UserPrincipal;
+import com.eventitta.api.auth.security.UserPrincipal;
 import com.eventitta.domain.user.api.internal.facade.UserInternalFacade;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
