+++ eventitta-api/src/main/java/com/eventitta/api/auth/jwt/JwtAccessDeniedHandler.java
@@ -1,6 +1,6 @@
 package com.eventitta.api.auth.jwt;
 
-import com.eventitta.api.auth.jwt.service.UserInfoService;
+import com.eventitta.api.auth.jwt.UserInfoService;
 import com.eventitta.domain.common.exception.CommonErrorCode;
 import com.eventitta.api.common.response.ApiErrorResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
