+++ eventitta-api/src/main/java/com/eventitta/api/auth/oauth/kakao/KakaoWebProperties.java
@@ -1,4 +1,4 @@
-package com.eventitta.api.auth.properties;
+package com.eventitta.api.auth.oauth.kakao;
 
 import lombok.Getter;
 import lombok.Setter;
