+++ eventitta-api/src/test/java/com/eventitta/api/auth/security/AuthenticationManagerCredentialAuthenticatorTest.java
@@ -1,6 +1,6 @@
 package com.eventitta.api.auth.security;
 
-import com.eventitta.api.auth.domain.UserPrincipal;
+import com.eventitta.api.auth.security.UserPrincipal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
