+++ eventitta-api/src/test/java/com/eventitta/api/auth/session/ClientSessionMetadataResolverTest.java
@@ -1,4 +1,4 @@
-package com.eventitta.api.auth.web;
+package com.eventitta.api.auth.session;
 
 import com.eventitta.domain.auth.service.dto.ClientSessionMetadata;
 import org.junit.jupiter.api.DisplayName;
