+++ eventitta-api/src/test/java/com/eventitta/api/user/controller/UserControllerSocialLinkTest.java
@@ -1,18 +1,18 @@
 package com.eventitta.api.user.controller;
 
-import com.eventitta.api.auth.config.SecurityConfig;
-import com.eventitta.api.auth.constants.AuthConstants;
+import com.eventitta.api.auth.SecurityConfig;
+import com.eventitta.api.auth.AuthConstants;
 import com.eventitta.api.auth.controller.request.SocialLoginRequest;
-import com.eventitta.api.auth.domain.UserPrincipal;
+import com.eventitta.api.auth.security.UserPrincipal;
 import com.eventitta.api.auth.jwt.JwtAccessDeniedHandler;
 import com.eventitta.api.auth.jwt.JwtAuthenticationEntryPoint;
 import com.eventitta.api.auth.jwt.JwtTokenProvider;
-import com.eventitta.api.auth.jwt.service.CustomUserDetailsService;
-import com.eventitta.api.auth.jwt.service.UserInfoService;
-import com.eventitta.api.auth.mapper.AuthMapperImpl;
-import com.eventitta.api.auth.properties.SecurityCorsProperties;
-import com.eventitta.api.auth.web.CookieManager;
-import com.eventitta.api.auth.web.KakaoAuthorizationSupport;
+import com.eventitta.api.auth.jwt.CustomUserDetailsService;
+import com.eventitta.api.auth.jwt.UserInfoService;
+import com.eventitta.api.auth.controller.AuthMapperImpl;
+import com.eventitta.api.auth.security.SecurityCorsProperties;
+import com.eventitta.api.auth.cookie.CookieManager;
+import com.eventitta.api.auth.oauth.kakao.KakaoAuthorizationSupport;
 import com.eventitta.api.user.mapper.UserMapper;
 import com.eventitta.domain.auth.exception.AuthErrorCode;
 import com.eventitta.domain.auth.repository.RefreshTokenRepository;
